api/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
python-dotenv==1.0.0
pydub==0.25.1
requests==2.31.0
pytz==2024.2
elevenlabs==0.2.27
gunicorn==20.1.0
Werkzeug==3.0.1